client/requirements.txt
# pip3 install -r requirements.txt

Flask
python-dotenv
matplotlib
mysql-connector
pytz
requests
web3
